requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.10.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pyjwt==2.8.0
python-dotenv==1.0.1
jinja2==3.1.3
xhtml2pdf==0.2.16
resend==0.8.0
python-multipart==0.0.9
email-validator
gunicorn==21.2.0
requests==2.31.0
python-bidi==0.4.2
pandas==2.2.1
openpyxl==3.1.2
apscheduler==3.10.4
numpy
aiofiles
pytz
python-dateutil==2.9.0